# Optimize `permutations` with low effort (no change in algorithm)